# backend/swagger-ui/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Baustellenverwaltung",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/constructionSite/{constructionSiteId}/note": {
      "get": {
        "tags": [
          "note"
        ],
        "summary": "Returns all notes for specified construction sites",
        "description": "Returns a array of notes.",
        "operationId": "getConstructionSiteNotes",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "maximum": 10,
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Note"
                  },
                  "x-content-type": "application/json"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "Note"
      },
      "post": {
        "tags": [
          "note"
        ],
        "summary": "Add note to constructionSite",
        "operationId": "addNote",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "maximum": 10,
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Note object that should be added to constructionSite",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Note"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "description": "Invalid properties on Task",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "Note"
      }
    },
    "/constructionSite/{constructionSiteId}/note{noteId}": {
      "put": {
        "tags": [
          "note"
        ],
        "summary": "Update note by ID",
        "description": "Update Note",
        "operationId": "updateNote",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "noteId",
            "in": "path",
            "description": "ID of note that should be updated",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "description": "Note object that should be added to constructionSite",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Note"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "404": {
            "description": "Note not found",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "Note"
      },
      "delete": {
        "tags": [
          "note"
        ],
        "summary": "Delete note by ID",
        "description": "Deletes note from construction Site",
        "operationId": "deleteNote",
        "parameters": [
          {
            "name": "constructionSiteId",
            "in": "path",
            "description": "ID of constructionSite",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "noteId",
            "in": "path",
            "description": "ID of note that needs to be deleted",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted",
            "content": {}
          },
          "400": {
            "description": "Invalid parameters",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          },
          "404": {
            "description": "Note not found",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-swagger-router-controller": "Note"
      }
    },
...
  },
  "components": {
    "schemas": {
      "Task": {
        "required": [
          "assignedTo",
          "priority",
          "status",
          "title",
          "constructionSiteId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Task_createdBy"
          },
          "modifiedAt": {
            "type": "string",
            "example": "{}"
          },
          "createdAt": {
            "type": "string",
            "example": "{}"
          },
          "constructionSiteId": {
            "type": "number",
            "example": 0
          },
          "priority": {
            "type": "string",
            "enum": [
              "high",
              "normal",
              "low"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in_progress",
              "completed"
            ]
          },
          "assignedTo": {
            "$ref": "#/components/schemas/Task_createdBy"
          }
        },
        "example": {
          "createdAt": "{}",
          "title": "title",
          "createdBy": {
            "user": "name",
            "userId": "userId"
          },
          "modifiedAt": "{}",
          "id": 0,
          "text": "text",
          "priority": "high",
          "assignedTo": "{}",
          "status": "open"
        },
        "xml": {
          "name": "Task"
        }
      },
...
      "Note": {
        "required": [
          "text",
          "title",
          "constructionSiteId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "createdBy": {
            "$ref": "#/components/schemas/Task_createdBy"
          },
          "constructionSiteId": {
            "type": "number",
            "example": 0
          },
          "modifiedAt": {
            "type": "string"
          }
        },
        "example": {
          "title": "title",
          "createdBy": {
            "user": "name",
            "userId": "userId"
          },
          "modifiedAt": "modifiedAt",
          "id": 0,
          "text": "text"
        },
        "xml": {
          "name": "Note"
        }
      },
...
      "Task_createdBy": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "description": "name of user"
          },
          "userId": {
            "type": "string",
            "description": "userId reference"
          }
        },
        "example": {
          "name": "name",
          "userId": "userId"
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}